+++ lib/ace/tcp/governor/supervisor.ex
@@ -9,12 +9,14 @@ defmodule Ace.TCP.Governor.Supervisor do
   @doc """
   Starts a supervised pool of governor processes.
   """
-  def start_link(server_supervisor, listen_socket, acceptors, sup_opts \\ []) do
-    Supervisor.start_link(__MODULE__, {server_supervisor, listen_socket, acceptors}, sup_opts)
+  @spec start_link(pid, :inet.socket, non_neg_integer) :: {:ok, pid}
+  def start_link(server_supervisor, listen_socket, acceptors) do
+    Supervisor.start_link(__MODULE__, {server_supervisor, listen_socket, acceptors}, [])
   end
 
   ## SERVER CALLBACKS
 
+  @doc false
   def init({server_supervisor, listen_socket, acceptors}) do
     # To speed up a server multiple process can be listening for a connection simultaneously.
     # In this case n Governors will start n Servers listening before a single connection is received.
@@ -22,6 +24,7 @@ defmodule Ace.TCP.Governor.Supervisor do
       worker(Ace.TCP.Governor, [listen_socket, server_supervisor], id: "#{i}")
     end
 
+    # The number of governors should be kept constent, for each governor that crashes a replacement should be started.
     supervise(children, strategy: :one_for_one)
   end
 end