+++ lib/ace/tcp/server.ex
@@ -29,7 +29,85 @@ defmodule Ace.TCP.Server do
     end
   end
   ```
+
+  See the README.md for a complete overview on how to make a server available.
   """
+
+  @typedoc """
+  Information about the servers connection to the client
+  """
+  @type connection :: %{peer: {:inet.ip_address, :inet.port_number}}
+
+  @typedoc """
+  The current state of an individual server process.
+  """
+  @type state :: term
+
+  @typedoc """
+  The configuration used to start each server.
+
+  A server configuration consists of behaviour, the `module`, and state.
+  The module should implement the `Ace.TCP.Server` behaviour.
+  Any value can be passed as the state.
+  """
+  @type app :: {module, state}
+
+  @doc """
+  Invoked when a new client connects.
+  `accept/2` will block until a client connects and the server has initialised
+
+  The `state` is the second element in the `app` tuple that was used to start the endpoint.
+
+  Returning `{:nosend, state}` will setup a new server with internal `state`.
+  The state is perserved in the process loop and passed as the second option to subsequent callbacks.
+
+  Returning `{:nosend, state, timeout}` is the same as `{:send, state}`.
+  In addition `handle_info(:timeout, state)` will be called after `timeout` milliseconds, if no messages are received in that interval.
+
+  Returning `{:send, message, state}` or `{:send, message, state, timeout}` is similar to their `:nosend` counterparts,
+  except the `message` is sent as the first communication to the client.
+
+  Returning `{:close, state}` will shutdown the server without any messages being sent or recieved
+  """
+  @callback init(connection, state) ::
+    {:send, iodata, state} |
+    {:send, iodata, state, timeout} |
+    {:nosend, state} |
+    {:nosend, state, timeout} |
+    {:close, state}
+
+  @doc """
+  Every packet recieved from the client invokes this callback.
+
+  The return actions are the same as for the `init/2` callback
+
+  *No additional packets will be taken from the socket until this callback returns*
+  """
+  @callback handle_packet(binary, state) ::
+    {:send, term, state} |
+    {:send, term, state, timeout} |
+    {:nosend, state} |
+    {:nosend, state, timeout} |
+    {:close, state}
+
+  @doc """
+  Every erlang message recieved by the server invokes this callback.
+
+  The return actions are the same as for the `init/2` callback
+  """
+  @callback handle_info(term, state) ::
+    {:send, term, state} |
+    {:send, term, state, timeout} |
+    {:nosend, state} |
+    {:nosend, state, timeout} |
+    {:close, state}
+
+  @doc """
+  Called whenever the connection is terminated.
+  """
+  # All normal sockets should be closed by the client?
+  @callback terminate(reason, state) :: term when
+    reason: term
 
   # Use OTP behaviour so the server can be added to a supervision tree.
   use GenServer
@@ -51,8 +129,9 @@ defmodule Ace.TCP.Server do
   The server process is returned immediatly.
   This is allow a supervisor to start several servers without waiting for connections.
 
-  A provisioned server will remain in an awaiting state untill accept is called.
+  A provisioned server will remain in an awaiting state until accept is called.
   """
+  @spec start_link(app) :: GenServer.on_start
   def start_link(app) do
     GenServer.start_link(__MODULE__, {:app, app}, [])
   end
@@ -63,6 +142,8 @@ defmodule Ace.TCP.Server do
   Accept can only be called once for each server.
   After a connection has been closed the server will terminate.
   """
+  @spec accept(server, :inet.socket) :: :ok when
+    server: pid
   def accept(server, listen_socket) do
     GenServer.call(server, {:accept, listen_socket}, :infinity)
   end