+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   Passed directly to GenServer so integer and `:hibernate` responses are supported.
 - Support closing the connection from the server side.
   It is best to reserver server side closing for misbehaving connections.
+- Added callbacks to `Ace.TCP.Server` so that is can be included as a behaviour. 
 
 ## [0.6.3](https://github.com/CrowdHailer/Ace/tree/0.6.3) - 2016-10-24